chore(zbugs): move `schema.ts` to the root
To help convey it is shared between client and zero-cache.
