+++ apps/zbugs/src/hooks/use-zero.ts
@@ -1,3 +1,3 @@
 import {createUseZero} from '@rocicorp/zero/react';
-import type {Schema} from '../domain/schema.js';
+import type {Schema} from '../../schema.js';
 export const useZero = createUseZero<Schema>();
