+++ apps/zbugs/tsconfig.node.json
@@ -1,9 +1,4 @@
 {
   "extends": "../../tsconfig.json",
-  "include": [
-    "vite.config.ts",
-    "zero.config.ts",
-    "src/domain/schema.ts",
-    "api/index.ts"
-  ]
+  "include": ["vite.config.ts", "zero.config.ts", "schema.ts", "api/index.ts"]
 }
